LoudFundLogos
Here are two LoudFund logos. Austin will send you over the verbiage for
the Ad. You can add this ad now and have the button linked to
loudfund.com

Once we go official with Loudfund,  the button will need to link to
Loudfund.com/reubensbell .  Ill let know the verbiage and when to pull
the trigger asap.